+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix issue with plotting Schmid factor maps
 - Fix bug with maps `component` not updating after masking
 - Store EDX data as a dictionary called `EDX` accessible from the EBSD `Map` object
+- Add option to change IPF map background colour
 
 
 ## 0.93.2 (16-04-2021)